# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Template API",
        "contact": {
            "name": "Development Team",
            "email": "collinsuchinaka@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://api.template.com",
            "description": "Live OpenApi Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login a user and obtain token",
                "operationId": "/login",
                "parameters": [
                    {
                        "name": "email_or_username",
                        "in": "query",
                        "description": "The login parameters should be in a post request, email or username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Some optional other parameter",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a success response and an API token",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Bad request. When the merchant does not exist."
                    },
                    "422": {
                        "description": "Error: Bad request. When the merchant name(account_id) is not a word or mis-spelt."
                    },
                    "401": {
                        "description": "Error: Bad request. When the user does not exist."
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registers a user",
                "operationId": "/register/user",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "first Name of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last Name of the User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username of the user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password Confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Creates a new user, returns a success response with an API token",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/password/forgot": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send token to user's email for verification",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Bad request. When required parameters does not exist"
                    }
                }
            }
        },
        "/forgot/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset user's password using token sent to email",
                "operationId": "/forgot/reset",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token sent to user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "New password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a success response and an API token",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Bad request. When the merchant name(account_id) is not a word or mis-spelt."
                    },
                    "401": {
                        "description": "Error: Bad request. When the user does not exist."
                    }
                }
            }
        },
        "/": {
            "get": {
                "tags": [
                    "Extras"
                ],
                "summary": "Ping",
                "description": "For Heath check: perfect for testing service health in production",
                "operationId": "App\\Http\\Controllers\\ExampleController::ping",
                "responses": {
                    "200": {
                        "description": "Service is active",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/posts": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a post",
                "operationId": "/posts",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of the post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description of the post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Creates a new user, returns a success response with an API token",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Everything about authentication"
        },
        {
            "name": "User",
            "description": "User details"
        },
        {
            "name": "Extras",
            "description": "Uncategorized"
        },
        {
            "name": "Admin",
            "description": "Admin routes (Not in use)"
        }
    ]
}